requirements.txt
pyTelegramBotAPI
airtable-python-wrapper
uWSGI==2.0.18